rand_distr: add module zeta
